Update SBEMimage_DMcom_GMS3.s
I was trying to execute SBEMimage_DMcom_GMS3.s in DigitalMicrograph and received errrors at line 635 about expecting ')'. The parentheses looked fine, but I noticed the or in the while statement didn't match the || in other conditional statements in the code. After I changed that, timeElapsed and timeElapsed2 where not declared (time_elapsed and time_elapsed2 are), so I declared them.