Reset template_grid_index and template_ov_index to 0 when selected grid/OV is deleted
This fixes a bug that occurred when the user had selected a grid or OV in Main Controls, then deleted that grid/OV and tried to draw a new one.
Note: Renamed current_ov and current_grid to template_ov_index and template_grid_index, which reflects their purpose better.